rain and snow effects from wip sdl mod
An added enhancement can be made to this to add zbuffer if anyone asks
for it. This only matters if you want sprites and snow/rain to be
layered along the viewers depth axis properly.

This rain/snow system is aimed to be resolution independent to some
extent.

An additional rain splash effect is applied on the ground where rain
particles land.

Snow particles respond to a sine-wave wind force which moves in a random
direction over time.