fix for wasserstein II crash
With some luck I reproduced a crash on E1M3 of Wasserstein II. It occurred before reaching the spear. I opened the last pushwall before the spear and then pressed F8 to save the game. The game then crashed during the save process and I was able to inspect a full trace within the debugger.

I have converted the weapon slot states into a fixed-size array and added some GC::Mark usage to be consistent with the other weapon pointers such as ReadyWeapon and PendingWeapon.

I could not reproduce the same crash in E1M3 after adding these changes.