fix broken sectors
Some maps with a mixture of indoor and outdoor areas can repro a broken
sector effect. The effect is visible on the ceiling which should
otherwise appear clear and show the sky.

This fix should enable the use of built-in rain effect combined with
indoor and outdoor areas in a single map.